DS-1289 Rearrange Integration Test Functions and Folder Structure (#813)
# Test Branch Pull Request

This PR re-arranges the test suite folder structure and moves some test
functions to more findable places.
